refactor: 다크모드 스토리지 싱크 기능 로직 변경
