+++ src/App.tsx
@@ -1,22 +1,21 @@
 import "./index.css";
 import { Route, Routes } from "react-router-dom";
 import { useGrain } from "./utils/hooks/useGrain";
-import { useDarkModeWithStorage } from "./utils/hooks/useDarkModeWithStorage";
-import { useDarkModeState } from "./utils/hooks/useDarkModeState";
+import { useDarkMode } from "./utils/hooks/useDarkMode";
 
 import Main from "./Pages/Main";
 import About from "./Pages/About";
 import Projects from "./Pages/Projects";
 import Contacts from "./Pages/Contacts";
 import Skills from "./Pages/Skills";
+import DarkModeWrapper from "./utils/DarkModeWrapper";
 
 function App() {
-  const darkMode = useDarkModeState();
+  const darkMode = useDarkMode();
   const canvasRef = useGrain(darkMode);
-  useDarkModeWithStorage();
 
   return (
-    <>
+    <DarkModeWrapper>
       <canvas
         ref={canvasRef}
         className="absolute top-0 left-0 w-full h-full z-10 opacity-5 pointer-events-none"
@@ -28,7 +27,7 @@ function App() {
         <Route path="/projects" element={<Projects />} />
         <Route path="/contacts" element={<Contacts />} />
       </Routes>
-    </>
+    </DarkModeWrapper>
   );
 }
